Created DemographicsDisplay component and updated day-of demo page.
