+++ src/applications/check-in/day-of/pages/Demographics.jsx
@@ -6,8 +6,8 @@ import recordEvent from 'platform/monitoring/record-event';
 import { goToNextPage, URLS } from '../utils/navigation';
 import BackToHome from '../components/BackToHome';
 import Footer from '../components/Footer';
-import DemographicItem from '../../components/DemographicItem';
 import { seeStaffMessageUpdated } from '../actions';
+import DemographicsDisplay from '../../components/pages/DemographicsDisplay';
 
 const Demographics = props => {
   const {
@@ -57,14 +57,6 @@ const Demographics = props => {
     [router, updateSeeStaffMessage],
   );
 
-  const demographicFields = [
-    { title: 'Mailing address', key: 'mailingAddress' },
-    { title: 'Home address', key: 'homeAddress' },
-    { title: 'Home phone', key: 'homePhone' },
-    { title: 'Mobile phone', key: 'mobilePhone' },
-    { title: 'Work phone', key: 'workPhone' },
-    { title: 'Email address', key: 'emailAddress' },
-  ];
   if (isLoading) {
     return <LoadingIndicator message={'Loading your appointments for today'} />;
   } else if (!demographics) {
@@ -72,50 +64,15 @@ const Demographics = props => {
     return <></>;
   } else {
     return (
-      <div className="vads-l-grid-container vads-u-padding-bottom--6 vads-u-padding-top--2 check-in-demographics">
-        <h1>Is this your current contact information?</h1>
-        <p className="vads-u-font-family--serif">
-          We can better follow up with you after your appointment when we have
-          your current information.
-        </p>
-        <div className="vads-u-border-color--primary vads-u-border-left--5px vads-u-margin-left--0p5 vads-u-padding-left--2">
-          <dl>
-            {demographicFields.map(demographicField => (
-              <React.Fragment key={demographicField.key}>
-                <dt className="vads-u-font-size--h3 vads-u-font-family--serif">
-                  {demographicField.title}
-                </dt>
-                <dd>
-                  {demographicField.key in demographics &&
-                  demographics[demographicField.key] ? (
-                    <DemographicItem
-                      demographic={demographics[demographicField.key]}
-                    />
-                  ) : (
-                    'Not available'
-                  )}
-                </dd>
-              </React.Fragment>
-            ))}
-          </dl>
-        </div>
-        <button
-          onClick={() => yesClick()}
-          className={'usa-button-secondary'}
-          data-testid="yes-button"
-        >
-          Yes
-        </button>
-        <button
-          onClick={() => noClick()}
-          className="usa-button-secondary vads-u-margin-top--2"
-          data-testid="no-button"
-        >
-          No
-        </button>
-        <Footer />
+      <>
+        <DemographicsDisplay
+          demographics={demographics}
+          yesAction={yesClick}
+          noAction={noClick}
+          Footer={Footer}
+        />
         <BackToHome />
-      </div>
+      </>
     );
   }
 };
